requirements.txt
# 🚀 REQUIREMENTS.TXT MAIS ATUALIZADO - 2024
# ✅ Versões testadas e funcionando no sistema
# 📅 Atualizado em: 24/06/2025
# 🎯 Status: Todas as dependências resolvidas

# ==========================================
# CORE WEB FRAMEWORK
# ==========================================
flask==2.3.3
flask-cors==4.0.0
werkzeug>=2.3.7

# ==========================================
# AWS SDK (Versões mais recentes)
# ==========================================
boto3==1.38.42
botocore==1.38.42

# ==========================================
# DOCUMENT PROCESSING
# ==========================================
PyPDF2==3.0.1
python-docx==0.8.11
lxml>=5.4.0

# ==========================================
# LANGCHAIN (Versões compatíveis - SEM CONFLITOS)
# ==========================================
langchain==0.3.26
langchain-aws==0.2.27
langchain-community==0.3.26
langchain-core==0.3.66
langchain-text-splitters==0.3.8

# ==========================================
# VECTOR STORE & ML
# ==========================================
faiss-cpu==1.7.4
numpy==1.26.4

# ==========================================
# UTILITIES
# ==========================================
python-dotenv==1.0.0
requests==2.32.4
typing-extensions>=4.7.0

# ==========================================
# WEB INTERFACE
# ==========================================
streamlit==1.46.0
pandas==2.3.0
plotly>=5.17.0

# ==========================================
# DEPENDENCIES AUTOMÁTICAS (Instaladas junto)
# ==========================================
# pydantic>=2.7.4
# SQLAlchemy>=2.0.0
# aiohttp>=3.8.3
# httpx>=0.23.0
# tenacity>=8.1.0
# jsonpatch>=1.33
# packaging>=23.2

# ==========================================
# OPCIONAL - SENTENCE TRANSFORMERS
# ==========================================
# sentence-transformers>=2.2.0  # Descomente se precisar

# ==========================================
# NOTAS DE INSTALAÇÃO
# ==========================================
# Para instalar:
# pip install -r requirements_2024_updated.txt
#
# Para instalação limpa:
# pip cache purge
# pip install -r requirements_2024_updated.txt
#
# Para resolver conflitos:
# python3 install_clean.py
